Add debounce configuration values
I've been experiencing key chatter/bounce issues with my Advantage360, and upon inspection it seems that the press and release debounce config values are both set to `-1` according to the build logs. Unless I'm missing configuration somewhere else, this means the firmware is effectively doing no debouncing.

Here I've set the values to the [ZMK defaults](https://zmk.dev/docs/features/debouncing), which are 5ms for both. On my own board this has removed the key bounce issue without adding any perceptible delay.

It's possible that this is just an issue with my board, but if debounce really is disabled then I'd expect _most_ people to experience this same issue to a certain degree.