Fix Makefile errors that prevent builds on macOS
Back in January 2023, some changes were made to the Makefile to drop
some flags that aren't available in macOS builds, that were causing
those builds to fail.

Unfortunately, these changes had a few small syntax issues that prevent
the conditional checks for "Darwin" from ever being true. See commits
0e19696, d011eea, 0737d53.

Specifically, a trailing comment after a variable assignment with spaces
before the comment character actually has the effect of adding those
trailing spaces to the variable being assigned (`detected_OS`). The
subsequent line attempts to strip this spaces, but inadvertently adds
them back with another trailing comment. Further, the quotes around
these assignments added in commit 0737d53 also end up verbatim in the
variable's value. The consequence of this the comparison actually ends up
comparing `"Darwin" ` with `Darwin`, so the check never succeeds and the
wrong flags are used. You can confirm this by adding

    echo .$(detected_OS).

to the `all` target and running make. You'll see the quotes and extra
spaces between the dots when the echo command is written to the console
(the quotes are eaten by the shell when echo is executed).

The fix is to remove the unnecessary quotes, the comments, and the now
unnecessary call to `strip`. I went a bit further and removed `detected_OS`
altogether since it's only used in one place.

I tested these changes on macOS 14.4 and Ubuntu Linux 22.04.4 LTS